Cleanup for Calendar Object & Methods
Decided what was really necessary for Calendar object. It's purpose was to supplement Date object, not replace it.

The Date constructor didn't return, or even store Months in it. It would return the index of the Month. The Calendar does two things well so far, stores the months of a year and the # of days they have, and allows you to 'get' a month using a shorthand for the month name (ie: Oct).