Code Maintanence
Made Calendar object a constant.

Months are stored with capitals because their words are things, not regular words. They should be stored as their words original meaning.

Changed Month to Months to reflect the change of MonthLength to Days. The data are plurals of the thing so their containers should be named properly.

Dates of holiday are grouped as a nested array.