Uncoupling reliance on the rst property
Removing the reliance of the rst property through send() method and acting on stored rst in internal property.

Instead, query result objects should be stored in variables on the pages they are used with identifiers unique to their case. Attempts to keep stored data minimal and acting or overwriting data minimal.

Use of structured queries available such as select and insert. Statement structure is specific. Pass in variables for user input. For any variables that are optional such as conditions and formatting, function determines whether keywords are added or not.