getISO regex functions
Functions which return formatted dates and times from the universal time standard (ISO) string of the JS Date object.

Formats the ISO string using regex.

It was discussed on 30-seconds-of-JavaScript that functions formatting date times should act on a universal standard passed into the function, rather than instantiating a Date object within the function and using a given time to construct an instance from. The latter returns inconsistent results due to the time zone effects of hardware, or whether the user passes through their time zone into the function.