JSON Calendars
Placed calendar information for older/foreign calendars into independent json files. The Calendar interface is being rewritten to accept a json collection such as these which it can work with to return monthnames using their index, property name, or their value which is the days in that month. This solution is more flexible as it works with any collection given to it and assumes no standard calendar as they are societaly dependent.